use size_t and not int as the prototype says so
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@3614 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/common/sprintf.c
@@ -138,7 +138,7 @@ static int sprfunc(void *ptr, unsigned char letter)
 }
 
 
-int snprintf(char *buf, int size, const char *fmt, ...)
+int snprintf(char *buf, size_t size, const char *fmt, ...)
 {
     bool ok;
     va_list ap;
